requirements.txt
pdfplumber>=0.10.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
click>=8.0.0
rich>=13.0.0
streamlit>=1.28.0